Move show.go content to config.go
